Initial working version
This has all of the templates and automation to build debs and rpms of the
release packages for our repositories. The rake task will build
puppetlabs-release for el (5,6), fedora (15,16), and debian/ubuntu using
rpmbuild, mock, and dpkg-buildpackage. The rpms are created in a layout
suitable for rsync to yum.puppetlabs.com. These packages all replace and
contain the devel packages that were specific to the release candidate
repositories.

It also has tasks to sign and ship debian and el packages to the rsync location
of choice. The el packages are signed using expect and debian packages are
signed using the -k flag for dpkg-buildpackage which can be combined with
gpg-agent.