(RE-4196) Move puppetlabs-release to vanagon
This commit moves puppetlabs-release package building to a vanagon style
project. This allows us to remove the old bespoke 'automation' in favor
of something maintainable. The repo definitions are moved from erb
templates to a tokenized text file that is updated using sed in the
build. Platform configs were imported from the puppet-agent repository.