(RE-4059) Update puppetlabs-release-pc1 to enable shipping
In order to actually ship the release package, PC1 must be added to the
project config, and a build_defaults.yaml file is needed with PC1 set
for the apt and yum repo_name. A rakefile is also required to access
some of the packaging tasks not yet available via vanagon.