(maint) Comment out the source repo
We are not shipping source packages yet for our puppet-agent releases.
As a result, running 'apt-get update' on a system where only
puppet-agent packages are available in the puppeltabs repo returns an
error. It's looking for the source repo, which isn't there. Since the
majority of new repos that we create do not yet have any sources, and
will not until puppetdb or puppetserver are shipped to them, we need to
deliver the repos with the source repo disabled.

If a user wants to access the source repos, all they have to do is
uncomment the line in the puppetlabs.list file provided by the release
package.