(maint) Bump version to 1.0.0
We really shouldn't be versioning this project from a git tag. There are
so few changes that happen, it makes a lot more sense to maintain a
single version. Most of the releases that happen will be for one
platform which does not already have this package available. Why do we
have to bump the version for one platform? Instead, if there are changes
that we need to make available on multiple platforms that already have
the releas package available, add in and bump the release field! Or bump
the version in this file.