(CPR-210) Allow puppetlabs-release-pc1 to not conflict with original
Prior to this commit, we unintentionally had the puppetlabs-relelase-pc1
conflict with puppetlabs-release, which is causing heartburn for some
folks. This fix simply changes the filename for the GPG key once
installed to include a -PC1 suffix.

The behavior now mimics the behavior on Debian, as in both release
packages can be installed.