(maint) Actually check the signature is correct
Prior to this commit, this changes how the automation validates
signatures on our rpm packages. Previously, this was just checking the
return status of a command, but that was effectively always returning
true. This commit upates the check to ensrue the key we signed the
package with (or attempted to sign the package with) is actually present
and returned when we run `rpm -Kv`