(maint) Remove signing automation
This commit removes all the signing automation contained in this repo.
It also adds in support for the packaging repo. This means we can take
advantage of the packaging repo to handle the signing and signature
verification process for these packages. The building and shipping
automation still lives in this repo.

This commit also updates the path where packages are dropped. It removes
the `rpm` directory from the package paths.