(maint) Ensure we provide only the new key
Prior to this change, we had been providing the new key with additional
signatures. This breaks yum on el5, unfortunately. This commit changes
the key so that we are using the most stripped down version of the key
possible. This means the key does not include any subkeys or additional
signatures.

The "canonical" way to do this is interactively editing the GPG key.
This generated key should have no subkeys and a single signature, its
self signature.

Though this is unnecessary on el based platforms beyond EL5, we are
using this minimal key on all platforms. This reduces complexity and
future maintenance costs.

Also, I have no idea why there are so few changes included with this
commit. The previous commit includes the new key with a subkey, which
this commit avoids somehow.