(maint) Set all release packages to same version
Prior to this commit, we were versioning the puppetlabs release packages
based on the platform version we were building them for. This was fine,
except it removed the ability to increment anything but the release
number. We are getting ready to make a large change to this package, but
in order to best to that, we should be able to increment the major or
minor version of the package. Unfortunately, due to the previous
versioning convention, we have to just increment the major version so
all the packages use the same one. Go us!