(maint) Add PC1 suffix to nightly key
Users in some cases need to install both the PC1 and "old school"
release packages on the same machine. The only issue is with conflicts
because both packages install the same keys. This is fine to have
duplicate keys installed, but rpm bails out when the files have the same
name. That's why we had to add the PC1 suffix to the puppet and
puppetlabs keys. We forgot to do the same for the nightly key. This
commit fixes that.