(RE-7959) Provide sles release packages
We recently added sles repos to yum.puppetlabs.com. In order to keep the
workflow for these new repos consistent with other rpm based platforms,
we need to provide release packages for these repos. However, because
sles and zypper like to make things interesting, there are a few tweaks
we need to make to these packages. Yay sles!

Signed-off-by: Melissa Stone <melissa@puppet.com>