(RE-8082) Fix repo definitions on sles
Prior to this commit, we had formatted the repo file to include two key
definitions for the yum repos. This is fine for most systems, but
sles/zypper cannot handle the format that we used. On these systems,
each new key definition needs to have a `=` before it to indicate it is
a new entry. This commit makes that change.