(maint) Only attempt to import the keys on sles 12
Prior to this commit, I was having issues installing the puppet release
packages on sles 11. The issue wasn't preventing the installation of
the packages, but it looks bad and it takes a long time to complete. It
only happens when we try to import the gpg keys during the postinstall
on sles 11.

```
warning: waiting for exclusive lock on /var/lib/rpm/Packages
error: cannot get exclusive lock on /var/lib/rpm/Packages
error: cannot open Packages index using db3 - Operation not permitted (1)
error: cannot open Packages database in /var/lib/rpm
error: /etc/pki/rpm-gpg/RPM-GPG-KEY-puppet-PC1: import failed.
```

Importing the keys was only a means to keep the sles release packages
in line with other platform release packages. When the keys are not
installed, the user must manually import them to make them available.
If the user does not manually import the keys, zypper will simply
complain the key to verify the signature is not available. Without this
change, zypper still complains about the unavailable key on sles 11.

This commit also updates the project release and all the component
versions as I had forgotten to do that last time.

Signed-off-by: Melissa Stone <melissa@puppet.com>