(CPR-383) Remove gpg key import on sles
So sles doesn't automatically import gpg keys when you use a repo for
the first time. To fix this, I included a postinst script to manually
import the keys for the user. This worked when I tested on sles 12, but
not on sles 11. It turns out the version of rpm on sles 11 isn't smart
enough to be able to call rpm within rpm, so it couldn't handle it. Sles
12 seemed to be smarter and able to handle this situation. However, it
seems some users are on older, less-updated versions of sles 12 and are
running into the issue we had seen on sles 11.

I give up, ya'll can just manually import the gpg keys yourselves.