(maint) Remove md5s from repo_definition config
This commit removes the md5sums from the repo_definition. Since the files we
are installing here are local to this repo, rather than being pulled from
upstream, we don't really need to include the md5. This will also allow us to
move forward in our efforts to consolidate the branches of this repo, since
each repo file would have a different md5sum, which would be difficult and/or
ugly to include in a shared config.