(maint) Update vanagon and packaging dependencies
This commit updates the `vanagon_location_for` method to be more generic and
makes use of this new `location_for` method to set gem versions for vanagon and
packaging. This commit also updates these dependencies to be loosely pinned so
that new non-major gem versions will automatically get pulled in.