(RE-13419) Make deb/sles/other structure more clear
In this commit, I needed to call out changes to sles with regard to
other yum-based distributions.

In doing so, I chopped up the repo_definition into a big case statement that
handles each platform individually and then moved the DSL
configuration to the very end.

The os_fixup hack was to defer the setting until the end without
frequent repetition of a long sed command. It could have been done
with a closure, most likely, but that seemed too hand-wavy.